Don't crash and burn when activating of the pcap device fails.
Fix up the Activate() function to return an std::optional<std::string>,
so that we can propagate error strings up to the caller. At the same
time, change the code to use a std::unique_ptr for the pcap_t, so that
Activate() becomes easier to understand and destruction does the right
thing implicitly.